Updated to a component-based structure, refactored utils, cleaned up index slightly.  About 30% of the way done with this.
