+++ index.html
@@ -146,7 +146,7 @@
             <a style="margin: 5px" href="https://hillsidehermitage.org">Hillside Hermitage Website</a>
         </div>
     </div>
-    <script type="module" src="index.js"></script>
+    <script type="module" src="/index.js"></script>
 </body>
 
 </html>
