+++ js/rawTextContent/forewordText.js
@@ -1,0 +1,15 @@
+export const forewordText = (`Terms and expressions of doctrinal and practical significance found in the early Suttas are sometimes misrepresented in existing translations. Unless readers habitually compare different translations to identify discrepancies, they may unknowingly incorporate these inaccuracies into their understanding of the texts. Good examples of this include: 
+  <br>
+  (1) The rendition of <em>mettā</em> as “loving-kindness” or even “love” that has become ingrained in contemporary Buddhism and had an impact on the general perception of what that practice entails. Passages like <a href="https://suttas.hillsidehermitage.org/?q=sn45.55">this</a> and <a href="https://suttas.hillsidehermitage.org/?q=an8.53">this</a>, as well as the etymology of the term, undermine that popular interpretation.
+  <br>
+  (2) The tendency to translate the term <em>yoniso manasikāra</em> along the lines of “appropriate” or “wise” attention, evidently assuming the literal meaning of <em>yoniso</em> to be unimportant. However, there is no reason to think that the Buddha didn’t intentionally opt for this peculiar expression to describe <a href="https://suttas.hillsidehermitage.org/?q=mn2#mn2:3.1-mn2:3.3">what is arguably the core element of the practice</a>, and <a href="https://suttas.hillsidehermitage.org/?q=sn45.55">leads to the acquisition of the Noble Eightfold Path</a>.
+  <br><br>
+ Although our modern retrospective understanding of Pāli will always be imperfect compared to that of ancient native speakers, it should always be firmly placed first, no matter how helpful or justified the translator believes a deviation to be. When a passage or Sutta does not align with one’s interpretation, it’s the interpretation that must be altered—from the bottom up if necessary—and not the rendition:
+  <br><br>
+  <em>It was, and is, my attitude towards the Suttas that, if I find anything in them that is against my own view, they are right, and I am wrong.</em>
+  <br>
+  —<a href=https://www.nanavira.org/post-sotapatti/1963/92-l-60-2-august-1963#n60-1>Ven. Ñāṇavīra Thera</a>
+  <br><br>
+  This site features adaptations of Bhikkhu Sujato's copyright-free translations, crafted with this context in mind. Elucidation beyond what is explicitly stated in the Pāli and in line with my view, references to relevant Suttas, and links to relevant discussions from the Hillside Hermitage YouTube channel can be found in the comments marked with a blue asterisk (<span class="comment">*</span>).
+  <br><br>
+  —<em>Bhikkhu Anīgha</em>`);
