+++ js/utils/misc/default_bookmark_dict.js
@@ -1,0 +1,1 @@
+export const DEFAULT_BOOKMARK_DICT = { bookmarks: { unlabeled: [] } };
