+++ js/utils/navigation/goBack.js
@@ -1,0 +1,3 @@
+export function goBack() {
+    window.history.back();
+}
