+++ js/utils/userActions/hideBookmarkButton.js
@@ -1,0 +1,6 @@
+export function hideBookmarkButton() {
+    let bookmarkButton = document.getElementById('bookmarkButton');
+    if (bookmarkButton) {
+      bookmarkButton.style.display = 'none';
+    }
+}
