+++ js/utils/userActions/hideCopyButton.js
@@ -1,0 +1,6 @@
+export function hideCopyButton() {
+    let copyButton = document.getElementById('copyButton');
+    if (copyButton) {
+      copyButton.style.display = 'none';
+    }
+  }
